requirements.txt
streamlit==1.32.0
torch==2.2.0
transformers==4.38.0
deep-translator==1.11.4
python-dotenv==1.0.0
openai==1.12.0
tqdm==4.66.0
pandas==2.2.0
numpy==1.24.3 